+++ src/Models/association/ProductAgeRange.ts
@@ -1,0 +1,7 @@
+import ProductAgeRangeInterface from '../../Interfaces/association/ProductAgeRange';
+
+export default class ProductAgeRange {
+    constructor(parameters: ProductAgeRangeInterface) {
+        Object.assign(this, parameters);
+    }
+}