+++ src/Models/association/ProductAssociation.ts
@@ -1,0 +1,7 @@
+import ProductAssociationInterface from '../../Interfaces/association/ProductAssociation';
+
+export default class ProductAssociation {
+    constructor(parameters: ProductAssociationInterface) {
+        Object.assign(this, parameters);
+    }
+}