+++ src/Models/association/ProductCategory.ts
@@ -1,0 +1,7 @@
+import ProductCategoryInterface from '../../Interfaces/association/ProductCategory';
+
+export default class ProductCategory {
+    constructor(parameters: ProductCategoryInterface) {
+        Object.assign(this, parameters);
+    }
+}