+++ src/Models/association/ProductCollection.ts
@@ -1,0 +1,7 @@
+import ProductCollectionInterface from '../../Interfaces/association/ProductCollection';
+
+export default class ProductCollection {
+    constructor(parameters: ProductCollectionInterface) {
+        Object.assign(this, parameters);
+    }
+}