+++ src/Models/association/ProductHazard.ts
@@ -1,0 +1,7 @@
+import ProductHazardInterface from '../../Interfaces/association/ProductHazard';
+
+export default class ProductHazard {
+    constructor(parameters: ProductHazardInterface) {
+        Object.assign(this, parameters);
+    }
+}