+++ src/Models/association/ProductManufacturer.ts
@@ -1,0 +1,7 @@
+import ProductManufacturerInterface from '../../Interfaces/association/ProductManufacturer';
+
+export default class ProductManufacturer {
+    constructor(parameters: ProductManufacturerInterface) {
+        Object.assign(this, parameters);
+    }
+}